+++ src/fields/emulated_fp/reduce.rs
@@ -1,4 +1,4 @@
-use super::{overhead, params::get_params, AllocatedNonNativeFieldVar};
+use super::{overhead, params::get_params, AllocatedEmulatedFpVar};
 use crate::{
     alloc::AllocVar,
     boolean::Boolean,
@@ -15,10 +15,7 @@ use ark_std::{cmp::min, marker::PhantomData, vec, vec::Vec};
 use num_bigint::BigUint;
 use num_integer::Integer;
 
-pub fn limbs_to_bigint<BaseField: PrimeField>(
-    bits_per_limb: usize,
-    limbs: &[BaseField],
-) -> BigUint {
+pub fn limbs_to_bigint<BaseF: PrimeField>(bits_per_limb: usize, limbs: &[BaseF]) -> BigUint {
     let mut val = BigUint::zero();
     let mut big_cur = BigUint::one();
     let two = BigUint::from(2u32);
@@ -37,16 +34,15 @@ pub fn limbs_to_bigint<BaseField: PrimeField>(
     val
 }
 
-pub fn bigint_to_basefield<BaseField: PrimeField>(bigint: &BigUint) -> BaseField {
-    let mut val = BaseField::zero();
-    let mut cur = BaseField::one();
+pub fn bigint_to_basefield<BaseF: PrimeField>(bigint: &BigUint) -> BaseF {
+    let mut val = BaseF::zero();
+    let mut cur = BaseF::one();
     let bytes = bigint.to_bytes_be();
 
-    let basefield_256 =
-        BaseField::from_bigint(<BaseField as PrimeField>::BigInt::from(256u64)).unwrap();
+    let basefield_256 = BaseF::from_bigint(<BaseF as PrimeField>::BigInt::from(256u64)).unwrap();
 
     for byte in bytes.iter().rev() {
-        let bytes_basefield = BaseField::from(*byte as u128);
+        let bytes_basefield = BaseF::from(*byte as u128);
         val += cur * bytes_basefield;
 
         cur *= &basefield_256;
@@ -56,32 +52,28 @@ pub fn bigint_to_basefield<BaseField: PrimeField>(bigint: &BigUint) -> BaseField
 }
 
 /// the collections of methods for reducing the presentations
-pub struct Reducer<TargetField: PrimeField, BaseField: PrimeField> {
-    pub target_phantom: PhantomData<TargetField>,
-    pub base_phantom: PhantomData<BaseField>,
+pub struct Reducer<TargetF: PrimeField, BaseF: PrimeField> {
+    pub target_phantom: PhantomData<TargetF>,
+    pub base_phantom: PhantomData<BaseF>,
 }
 
-impl<TargetField: PrimeField, BaseField: PrimeField> Reducer<TargetField, BaseField> {
-    /// convert limbs to bits (take at most `BaseField::MODULUS_BIT_SIZE as
+impl<TargetF: PrimeField, BaseF: PrimeField> Reducer<TargetF, BaseF> {
+    /// convert limbs to bits (take at most `BaseF::MODULUS_BIT_SIZE as
     /// usize - 1` bits) This implementation would be more efficient than
     /// the original `to_bits` or `to_non_unique_bits` since we enforce that
     /// some bits are always zero.
     #[tracing::instrument(target = "r1cs")]
-    pub fn limb_to_bits(
-        limb: &FpVar<BaseField>,
-        num_bits: usize,
-    ) -> R1CSResult<Vec<Boolean<BaseField>>> {
+    pub fn limb_to_bits(limb: &FpVar<BaseF>, num_bits: usize) -> R1CSResult<Vec<Boolean<BaseF>>> {
         let cs = limb.cs();
 
-        let num_bits = min(BaseField::MODULUS_BIT_SIZE as usize - 1, num_bits);
+        let num_bits = min(BaseF::MODULUS_BIT_SIZE as usize - 1, num_bits);
         let mut bits_considered = Vec::with_capacity(num_bits);
         let limb_value = limb.value().unwrap_or_default();
 
-        let num_bits_to_shave =
-            BaseField::BigInt::NUM_LIMBS * 64 - (BaseField::MODULUS_BIT_SIZE as usize);
+        let num_bits_to_shave = BaseF::BigInt::NUM_LIMBS * 64 - (BaseF::MODULUS_BIT_SIZE as usize);
 
         for b in BitIteratorBE::new(limb_value.into_bigint())
-            .skip(num_bits_to_shave + (BaseField::MODULUS_BIT_SIZE as usize - num_bits))
+            .skip(num_bits_to_shave + (BaseF::MODULUS_BIT_SIZE as usize - num_bits))
         {
             bits_considered.push(b);
         }
@@ -89,25 +81,24 @@ impl<TargetField: PrimeField, BaseField: PrimeField> Reducer<TargetField, BaseFi
         if cs == ConstraintSystemRef::None {
             let mut bits = vec![];
             for b in bits_considered {
-                bits.push(Boolean::<BaseField>::Constant(b));
+                bits.push(Boolean::<BaseF>::Constant(b));
             }
 
             Ok(bits)
         } else {
             let mut bits = vec![];
             for b in bits_considered {
-                bits.push(Boolean::<BaseField>::new_witness(
+                bits.push(Boolean::<BaseF>::new_witness(
                     ark_relations::ns!(cs, "bit"),
                     || Ok(b),
                 )?);
             }
 
-            let mut bit_sum = FpVar::<BaseField>::zero();
-            let mut coeff = BaseField::one();
+            let mut bit_sum = FpVar::<BaseF>::zero();
+            let mut coeff = BaseF::one();
 
             for bit in bits.iter().rev() {
-                bit_sum +=
-                    <FpVar<BaseField> as From<Boolean<BaseField>>>::from((*bit).clone()) * coeff;
+                bit_sum += <FpVar<BaseF> as From<Boolean<BaseF>>>::from((*bit).clone()) * coeff;
                 coeff.double_in_place();
             }
 
@@ -119,11 +110,10 @@ impl<TargetField: PrimeField, BaseField: PrimeField> Reducer<TargetField, BaseFi
 
     /// Reduction to the normal form
     #[tracing::instrument(target = "r1cs")]
-    pub fn reduce(elem: &mut AllocatedNonNativeFieldVar<TargetField, BaseField>) -> R1CSResult<()> {
-        let new_elem =
-            AllocatedNonNativeFieldVar::new_witness(ns!(elem.cs(), "normal_form"), || {
-                Ok(elem.value().unwrap_or_default())
-            })?;
+    pub fn reduce(elem: &mut AllocatedEmulatedFpVar<TargetF, BaseF>) -> R1CSResult<()> {
+        let new_elem = AllocatedEmulatedFpVar::new_witness(ns!(elem.cs(), "normal_form"), || {
+            Ok(elem.value().unwrap_or_default())
+        })?;
         elem.conditional_enforce_equal(&new_elem, &Boolean::TRUE)?;
         *elem = new_elem;
 
@@ -132,17 +122,15 @@ impl<TargetField: PrimeField, BaseField: PrimeField> Reducer<TargetField, BaseFi
 
     /// Reduction to be enforced after additions
     #[tracing::instrument(target = "r1cs")]
-    pub fn post_add_reduce(
-        elem: &mut AllocatedNonNativeFieldVar<TargetField, BaseField>,
-    ) -> R1CSResult<()> {
+    pub fn post_add_reduce(elem: &mut AllocatedEmulatedFpVar<TargetF, BaseF>) -> R1CSResult<()> {
         let params = get_params(
-            TargetField::MODULUS_BIT_SIZE as usize,
-            BaseField::MODULUS_BIT_SIZE as usize,
+            TargetF::MODULUS_BIT_SIZE as usize,
+            BaseF::MODULUS_BIT_SIZE as usize,
             elem.get_optimization_type(),
         );
-        let surfeit = overhead!(elem.num_of_additions_over_normal_form + BaseField::one()) + 1;
+        let surfeit = overhead!(elem.num_of_additions_over_normal_form + BaseF::one()) + 1;
 
-        if BaseField::MODULUS_BIT_SIZE as usize > 2 * params.bits_per_limb + surfeit + 1 {
+        if BaseF::MODULUS_BIT_SIZE as usize > 2 * params.bits_per_limb + surfeit + 1 {
             Ok(())
         } else {
             Self::reduce(elem)
@@ -153,8 +141,8 @@ impl<TargetField: PrimeField, BaseField: PrimeField> Reducer<TargetField, BaseFi
     /// way that allows efficient multiplication
     #[tracing::instrument(target = "r1cs")]
     pub fn pre_mul_reduce(
-        elem: &mut AllocatedNonNativeFieldVar<TargetField, BaseField>,
-        elem_other: &mut AllocatedNonNativeFieldVar<TargetField, BaseField>,
+        elem: &mut AllocatedEmulatedFpVar<TargetF, BaseF>,
+        elem_other: &mut AllocatedEmulatedFpVar<TargetF, BaseF>,
     ) -> R1CSResult<()> {
         assert_eq!(
             elem.get_optimization_type(),
@@ -162,30 +150,29 @@ impl<TargetField: PrimeField, BaseField: PrimeField> Reducer<TargetField, BaseFi
         );
 
         let params = get_params(
-            TargetField::MODULUS_BIT_SIZE as usize,
-            BaseField::MODULUS_BIT_SIZE as usize,
+            TargetF::MODULUS_BIT_SIZE as usize,
+            BaseF::MODULUS_BIT_SIZE as usize,
             elem.get_optimization_type(),
         );
 
         if 2 * params.bits_per_limb + ark_std::log2(params.num_limbs) as usize
-            > BaseField::MODULUS_BIT_SIZE as usize - 1
+            > BaseF::MODULUS_BIT_SIZE as usize - 1
         {
             panic!("The current limb parameters do not support multiplication.");
         }
 
         loop {
-            let prod_of_num_of_additions = (elem.num_of_additions_over_normal_form
-                + BaseField::one())
-                * (elem_other.num_of_additions_over_normal_form + BaseField::one());
+            let prod_of_num_of_additions = (elem.num_of_additions_over_normal_form + BaseF::one())
+                * (elem_other.num_of_additions_over_normal_form + BaseF::one());
             let overhead_limb = overhead!(prod_of_num_of_additions.mul(
-                &BaseField::from_bigint(<BaseField as PrimeField>::BigInt::from(
+                &BaseF::from_bigint(<BaseF as PrimeField>::BigInt::from(
                     (params.num_limbs) as u64
                 ))
                 .unwrap()
             ));
             let bits_per_mulresult_limb = 2 * (params.bits_per_limb + 1) + overhead_limb;
 
-            if bits_per_mulresult_limb < BaseField::MODULUS_BIT_SIZE as usize {
+            if bits_per_mulresult_limb < BaseF::MODULUS_BIT_SIZE as usize {
                 break;
             }
 
@@ -203,9 +190,7 @@ impl<TargetField: PrimeField, BaseField: PrimeField> Reducer<TargetField, BaseFi
 
     /// Reduction to the normal form
     #[tracing::instrument(target = "r1cs")]
-    pub fn pre_eq_reduce(
-        elem: &mut AllocatedNonNativeFieldVar<TargetField, BaseField>,
-    ) -> R1CSResult<()> {
+    pub fn pre_eq_reduce(elem: &mut AllocatedEmulatedFpVar<TargetF, BaseF>) -> R1CSResult<()> {
         if elem.is_in_the_normal_form {
             return Ok(());
         }
@@ -219,14 +204,14 @@ impl<TargetField: PrimeField, BaseField: PrimeField> Reducer<TargetField, BaseFi
         surfeit: usize,
         bits_per_limb: usize,
         shift_per_limb: usize,
-        left: &[FpVar<BaseField>],
-        right: &[FpVar<BaseField>],
+        left: &[FpVar<BaseF>],
+        right: &[FpVar<BaseF>],
     ) -> R1CSResult<()> {
         let cs = left.cs().or(right.cs());
-        let zero = FpVar::<BaseField>::zero();
+        let zero = FpVar::<BaseF>::zero();
 
-        let mut limb_pairs = Vec::<(FpVar<BaseField>, FpVar<BaseField>)>::new();
-        let num_limb_in_a_group = (BaseField::MODULUS_BIT_SIZE as usize
+        let mut limb_pairs = Vec::<(FpVar<BaseF>, FpVar<BaseF>)>::new();
+        let num_limb_in_a_group = (BaseF::MODULUS_BIT_SIZE as usize
             - 1
             - surfeit
             - 1
@@ -237,9 +222,9 @@ impl<TargetField: PrimeField, BaseField: PrimeField> Reducer<TargetField, BaseFi
 
         let shift_array = {
             let mut array = Vec::new();
-            let mut cur = BaseField::one().into_bigint();
+            let mut cur = BaseF::one().into_bigint();
             for _ in 0..num_limb_in_a_group {
-                array.push(BaseField::from_bigint(cur).unwrap());
+                array.push(BaseF::from_bigint(cur).unwrap());
                 cur <<= shift_per_limb as u32;
             }
 
@@ -252,7 +237,7 @@ impl<TargetField: PrimeField, BaseField: PrimeField> Reducer<TargetField, BaseFi
             limb_pairs.push((left_limb.clone(), right_limb.clone()));
         }
 
-        let mut groupped_limb_pairs = Vec::<(FpVar<BaseField>, FpVar<BaseField>, usize)>::new();
+        let mut groupped_limb_pairs = Vec::<(FpVar<BaseF>, FpVar<BaseF>, usize)>::new();
 
         for limb_pairs_in_a_group in limb_pairs.chunks(num_limb_in_a_group) {
             let mut left_total_limb = zero.clone();
@@ -275,19 +260,19 @@ impl<TargetField: PrimeField, BaseField: PrimeField> Reducer<TargetField, BaseFi
         // This part we mostly use the techniques in bellman-bignat
         // The following code is adapted from https://github.com/alex-ozdemir/bellman-bignat/blob/master/src/mp/bignat.rs#L567
         let mut carry_in = zero;
-        let mut carry_in_value = BaseField::zero();
+        let mut carry_in_value = BaseF::zero();
         let mut accumulated_extra = BigUint::zero();
         for (group_id, (left_total_limb, right_total_limb, num_limb_in_this_group)) in
             groupped_limb_pairs.iter().enumerate()
         {
-            let mut pad_limb_repr = BaseField::ONE.into_bigint();
+            let mut pad_limb_repr = BaseF::ONE.into_bigint();
 
             pad_limb_repr <<= (surfeit
                 + (bits_per_limb - shift_per_limb)
                 + shift_per_limb * num_limb_in_this_group
                 + 1
                 + 1) as u32;
-            let pad_limb = BaseField::from_bigint(pad_limb_repr).unwrap();
+            let pad_limb = BaseF::from_bigint(pad_limb_repr).unwrap();
 
             let left_total_limb_value = left_total_limb.value().unwrap_or_default();
             let right_total_limb_value = right_total_limb.value().unwrap_or_default();
@@ -298,7 +283,7 @@ impl<TargetField: PrimeField, BaseField: PrimeField> Reducer<TargetField, BaseFi
             let carry_repr =
                 carry_value.into_bigint() >> (shift_per_limb * num_limb_in_this_group) as u32;
 
-            carry_value = BaseField::from_bigint(carry_repr).unwrap();
+            carry_value = BaseF::from_bigint(carry_repr).unwrap();
 
             let carry = FpVar::new_witness(cs.clone(), || Ok(carry_value))?;
 
@@ -307,7 +292,7 @@ impl<TargetField: PrimeField, BaseField: PrimeField> Reducer<TargetField, BaseFi
             let (new_accumulated_extra, remainder) = accumulated_extra.div_rem(
                 &BigUint::from(2u64).pow((shift_per_limb * num_limb_in_this_group) as u32),
             );
-            let remainder_limb = bigint_to_basefield::<BaseField>(&remainder);
+            let remainder_limb = bigint_to_basefield::<BaseF>(&remainder);
 
             // Now check
             //      left_total_limb + pad_limb + carry_in - right_total_limb
@@ -316,21 +301,21 @@ impl<TargetField: PrimeField, BaseField: PrimeField> Reducer<TargetField, BaseFi
             let eqn_left = left_total_limb + pad_limb + &carry_in - right_total_limb;
 
             let eqn_right = &carry
-                * BaseField::from(2u64).pow(&[(shift_per_limb * num_limb_in_this_group) as u64])
+                * BaseF::from(2u64).pow(&[(shift_per_limb * num_limb_in_this_group) as u64])
                 + remainder_limb;
 
-            eqn_left.conditional_enforce_equal(&eqn_right, &Boolean::<BaseField>::TRUE)?;
+            eqn_left.conditional_enforce_equal(&eqn_right, &Boolean::<BaseF>::TRUE)?;
 
             accumulated_extra = new_accumulated_extra;
             carry_in = carry.clone();
             carry_in_value = carry_value;
 
             if group_id == groupped_limb_pairs.len() - 1 {
-                carry.enforce_equal(&FpVar::<BaseField>::Constant(bigint_to_basefield(
+                carry.enforce_equal(&FpVar::<BaseF>::Constant(bigint_to_basefield(
                     &accumulated_extra,
                 )))?;
             } else {
-                Reducer::<TargetField, BaseField>::limb_to_bits(&carry, surfeit + bits_per_limb)?;
+                Reducer::<TargetF, BaseF>::limb_to_bits(&carry, surfeit + bits_per_limb)?;
             }
         }